Moved export function to fetch_sample_resource.py and refactored it to give appropriate response on fetch status
